# examples/server/beego/basic/gen.go
// Code generated by mockzilla-codegen. DO NOT EDIT.

package basic

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strings"

	"github.com/beego/beego/v2/server/web"
	bcontext "github.com/beego/beego/v2/server/web/context"
	"github.com/mockzilla/mockzilla-codegen/pkg/runtime"
)

// Fails to compile when the runtime does not match the generator that wrote this file.
const _ = runtime.SupportsGeneratorV2

type Pet struct {
	ID   int    `json:"id"`
	Name string `json:"name"`
}

type ListPetsQuery struct {
	Limit *int `json:"limit,omitempty"`
}

type GetPetPathParams struct {
	ID int `json:"id"`
}

type DeletePetPathParams struct {
	ID int `json:"id"`
}

type ListPetsResponse200 []Pet

type CreatePetResponse201Headers struct {
	Location *string `json:"Location,omitempty"`
}

type PingResponse200 = string

// PetsInterface is what the generated handlers call. Implement it with the business logic.
type PetsInterface interface {
	// ListPets handles GET /pets.
	ListPets(ctx context.Context, opts *ListPetsServiceRequestOptions) (*ListPetsResponseData, error)
	// CreatePet handles POST /pets.
	CreatePet(ctx context.Context, opts *CreatePetServiceRequestOptions) (*CreatePetResponseData, error)
	// GetPet handles GET /pets/{id}.
	GetPet(ctx context.Context, opts *GetPetServiceRequestOptions) (*GetPetResponseData, error)
	// DeletePet handles DELETE /pets/{id}.
	DeletePet(ctx context.Context, opts *DeletePetServiceRequestOptions) (*DeletePetResponseData, error)
	// Ping handles GET /ping.
	Ping(ctx context.Context, opts *PingServiceRequestOptions) (*PingResponseData, error)
}

// ListPetsServiceRequestOptions is what ListPets receives.
type ListPetsServiceRequestOptions struct {
	Query      *ListPetsQuery
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *ListPetsServiceRequestOptions) Validate() error {
	return nil
}

// ListPetsResponseData is what ListPets returns.
type ListPetsResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewListPetsResponseData returns the 200 response with its application/json body.
func NewListPetsResponseData(body ListPetsResponse200) *ListPetsResponseData {
	return &ListPetsResponseData{Status: 200, Body: body, contentType: "application/json"}
}

// WithStatus sets the status code.
func (r *ListPetsResponseData) WithStatus(code int) *ListPetsResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *ListPetsResponseData) WithHeaders(h http.Header) *ListPetsResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *ListPetsResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *ListPetsResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *ListPetsResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *ListPetsResponseData) ContentType() string {
	return r.contentType
}

// CreatePetServiceRequestOptions is what CreatePet receives.
type CreatePetServiceRequestOptions struct {
	// Body sent as application/json.
	Body       *Pet
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *CreatePetServiceRequestOptions) Validate() error {
	return nil
}

// CreatePetResponseData is what CreatePet returns.
type CreatePetResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewCreatePetResponseData returns the 201 response with its application/json body.
func NewCreatePetResponseData(body *Pet) *CreatePetResponseData {
	return &CreatePetResponseData{Status: 201, Body: body, contentType: "application/json"}
}

// WithStatus sets the status code.
func (r *CreatePetResponseData) WithStatus(code int) *CreatePetResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *CreatePetResponseData) WithHeaders(h http.Header) *CreatePetResponseData {
	r.Headers = h
	return r
}

// WithTypedHeaders adds the headers the spec declares for status 201.
func (r *CreatePetResponseData) WithTypedHeaders(h CreatePetResponse201Headers) *CreatePetResponseData {
	r.Headers = runtime.Headers(r.Headers, h)
	return r
}

// StatusCode returns the status.
func (r *CreatePetResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *CreatePetResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *CreatePetResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *CreatePetResponseData) ContentType() string {
	return r.contentType
}

// GetPetServiceRequestOptions is what GetPet receives.
type GetPetServiceRequestOptions struct {
	PathParams *GetPetPathParams
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *GetPetServiceRequestOptions) Validate() error {
	return nil
}

// GetPetResponseData is what GetPet returns.
type GetPetResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewGetPetResponseData200 returns the 200 response with its application/json body.
func NewGetPetResponseData200(body *Pet) *GetPetResponseData {
	return &GetPetResponseData{Status: 200, Body: body, contentType: "application/json"}
}

// NewGetPetResponseData404 returns the 404 response.
func NewGetPetResponseData404() *GetPetResponseData {
	return &GetPetResponseData{Status: 404, contentType: ""}
}

// WithStatus sets the status code.
func (r *GetPetResponseData) WithStatus(code int) *GetPetResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *GetPetResponseData) WithHeaders(h http.Header) *GetPetResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *GetPetResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *GetPetResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *GetPetResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *GetPetResponseData) ContentType() string {
	return r.contentType
}

// DeletePetServiceRequestOptions is what DeletePet receives.
type DeletePetServiceRequestOptions struct {
	PathParams *DeletePetPathParams
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *DeletePetServiceRequestOptions) Validate() error {
	return nil
}

// DeletePetResponseData is what DeletePet returns.
type DeletePetResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewDeletePetResponseData returns the 204 response.
func NewDeletePetResponseData() *DeletePetResponseData {
	return &DeletePetResponseData{Status: 204, contentType: ""}
}

// WithStatus sets the status code.
func (r *DeletePetResponseData) WithStatus(code int) *DeletePetResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *DeletePetResponseData) WithHeaders(h http.Header) *DeletePetResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *DeletePetResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *DeletePetResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *DeletePetResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *DeletePetResponseData) ContentType() string {
	return r.contentType
}

// PingServiceRequestOptions is what Ping receives.
type PingServiceRequestOptions struct {
	RawRequest *http.Request
}

// Validate checks the parameters and the body against the constraints of the spec.
func (o *PingServiceRequestOptions) Validate() error {
	return nil
}

// PingResponseData is what Ping returns.
type PingResponseData struct {
	Status  int
	Headers http.Header
	Body    any

	contentType string
}

// NewPingResponseData returns the 200 response with its text/plain body.
func NewPingResponseData(body *PingResponse200) *PingResponseData {
	return &PingResponseData{Status: 200, Body: body, contentType: "text/plain"}
}

// WithStatus sets the status code.
func (r *PingResponseData) WithStatus(code int) *PingResponseData {
	r.Status = code
	return r
}

// WithHeaders sets the headers.
func (r *PingResponseData) WithHeaders(h http.Header) *PingResponseData {
	r.Headers = h
	return r
}

// StatusCode returns the status.
func (r *PingResponseData) StatusCode() int {
	return r.Status
}

// Header returns the headers.
func (r *PingResponseData) Header() http.Header {
	return r.Headers
}

// Payload returns the body.
func (r *PingResponseData) Payload() any {
	return r.Body
}

// ContentType returns the media type of the body, empty for the default of its Go type.
func (r *PingResponseData) ContentType() string {
	return r.contentType
}

// The error types the handlers use, as the runtime declares them.
type (
	ErrorKind           = runtime.ErrorKind
	HandlerError        = runtime.HandlerError
	ErrorHandler        = runtime.ErrorHandler
	ErrorHandlerFunc    = runtime.ErrorHandlerFunc
	DefaultErrorHandler = runtime.DefaultErrorHandler
)

// The kinds of HandlerError.
const (
	ErrorParse      = runtime.ErrorParse
	ErrorDecode     = runtime.ErrorDecode
	ErrorValidation = runtime.ErrorValidation
	ErrorService    = runtime.ErrorService
	ErrorResponse   = runtime.ErrorResponse
)

// ServerOptions is what the adapter and the router are set up with.
type ServerOptions struct {
	Router             any
	Middleware         []func(http.Handler) http.Handler
	ErrorHandler       runtime.ErrorHandler
	JSONDecoder        func(body io.Reader, dst any, isRequired bool) error
	MultipartMaxMemory int64
}

// ServerOption sets one field of ServerOptions.
type ServerOption func(*ServerOptions)

// NewServerOptions applies opts to the defaults.
func NewServerOptions(opts ...ServerOption) *ServerOptions {
	o := &ServerOptions{
		ErrorHandler:       runtime.DefaultErrorHandler{},
		JSONDecoder:        runtime.DecodeJSON,
		MultipartMaxMemory: 33554432,
	}
	for _, opt := range opts {
		opt(o)
	}
	return o
}

// WithMiddleware wraps the routes with mw, outermost first; on a new router, unknown paths too.
func WithMiddleware(mw ...func(http.Handler) http.Handler) ServerOption {
	return func(o *ServerOptions) {
		o.Middleware = append(o.Middleware, mw...)
	}
}

// WithErrorHandler sets what writes the response of a failed request.
func WithErrorHandler(h runtime.ErrorHandler) ServerOption {
	return func(o *ServerOptions) {
		o.ErrorHandler = h
	}
}

// WithJSONDecoder sets what reads JSON bodies. isRequired says whether an empty body is an error.
func WithJSONDecoder(decode func(body io.Reader, dst any, isRequired bool) error) ServerOption {
	return func(o *ServerOptions) {
		o.JSONDecoder = decode
	}
}

// WithMultipartMaxMemory sets how much of a multipart form stays in memory.
func WithMultipartMaxMemory(n int64) ServerOption {
	return func(o *ServerOptions) {
		o.MultipartMaxMemory = n
	}
}

// HTTPAdapter answers HTTP requests by calling the service.
type HTTPAdapter struct {
	svc  PetsInterface
	opts *ServerOptions
}

type responseData interface {
	StatusCode() int
	Header() http.Header
	Payload() any
	ContentType() string
}

// NewHTTPAdapter returns the adapter of svc.
func NewHTTPAdapter(svc PetsInterface, opts ...ServerOption) *HTTPAdapter {
	return &HTTPAdapter{svc: svc, opts: NewServerOptions(opts...)}
}

// ListPets handles GET /pets.
func (a *HTTPAdapter) ListPets(w http.ResponseWriter, r *http.Request) {
	opts := &ListPetsServiceRequestOptions{RawRequest: r}
	query := r.URL.Query()
	opts.Query = &ListPetsQuery{}
	if err := runtime.DecodeQuery(query, runtime.Param{Name: "limit", Style: runtime.StyleForm, IsExplode: true, IsRequired: false, IsJSON: false}, &opts.Query.Limit); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "ListPets", ParamName: "limit", ParamLocation: "query", Err: err})
		return
	}

	res, err := a.svc.ListPets(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "ListPets", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "ListPets", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "ListPets", res)
}

// CreatePet handles POST /pets.
func (a *HTTPAdapter) CreatePet(w http.ResponseWriter, r *http.Request) {
	opts := &CreatePetServiceRequestOptions{RawRequest: r}
	switch contentType := runtime.ContentType(r.Header); contentType {
	case "application/json":
		if err := a.opts.JSONDecoder(r.Body, &opts.Body, true); err != nil {
			a.failDecode(w, r, "CreatePet", err)
			return
		}
	case "":
		a.failDecode(w, r, "CreatePet", runtime.ErrBodyEmpty)
		return
	default:
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorDecode, OperationID: "CreatePet", Status: http.StatusUnsupportedMediaType, Err: runtime.ContentTypeError(contentType)})
		return
	}

	res, err := a.svc.CreatePet(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "CreatePet", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "CreatePet", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "CreatePet", res)
}

// GetPet handles GET /pets/{id}.
func (a *HTTPAdapter) GetPet(w http.ResponseWriter, r *http.Request) {
	opts := &GetPetServiceRequestOptions{RawRequest: r}
	opts.PathParams = &GetPetPathParams{}
	if err := runtime.DecodePath(r.PathValue("id"), runtime.Param{Name: "id", Style: runtime.StyleSimple, IsExplode: false, IsRequired: true, IsJSON: false}, &opts.PathParams.ID); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "GetPet", ParamName: "id", ParamLocation: "path", Err: err})
		return
	}

	res, err := a.svc.GetPet(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "GetPet", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "GetPet", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "GetPet", res)
}

// DeletePet handles DELETE /pets/{id}.
func (a *HTTPAdapter) DeletePet(w http.ResponseWriter, r *http.Request) {
	opts := &DeletePetServiceRequestOptions{RawRequest: r}
	opts.PathParams = &DeletePetPathParams{}
	if err := runtime.DecodePath(r.PathValue("id"), runtime.Param{Name: "id", Style: runtime.StyleSimple, IsExplode: false, IsRequired: true, IsJSON: false}, &opts.PathParams.ID); err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorParse, OperationID: "DeletePet", ParamName: "id", ParamLocation: "path", Err: err})
		return
	}

	res, err := a.svc.DeletePet(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "DeletePet", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "DeletePet", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "DeletePet", res)
}

// Ping handles GET /ping.
func (a *HTTPAdapter) Ping(w http.ResponseWriter, r *http.Request) {
	opts := &PingServiceRequestOptions{RawRequest: r}

	res, err := a.svc.Ping(r.Context(), opts)
	if err != nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "Ping", Err: err})
		return
	}
	if res == nil {
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: "Ping", Err: runtime.ErrNoResponse})
		return
	}
	a.write(w, r, "Ping", res)
}

func (a *HTTPAdapter) fail(w http.ResponseWriter, r *http.Request, err *runtime.HandlerError) {
	// A response that failed to write leaves its media type, which is not the error's.
	w.Header().Del("Content-Type")
	a.opts.ErrorHandler.HandleError(w, r, err.StatusCode(), err)
}

func (a *HTTPAdapter) failDecode(w http.ResponseWriter, r *http.Request, id string, err error) {
	a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorDecode, OperationID: id, Err: err})
}

func (a *HTTPAdapter) write(w http.ResponseWriter, r *http.Request, id string, res responseData) {
	if res.ContentType() != "" {
		w.Header().Set("Content-Type", res.ContentType())
	}
	err := runtime.Write(w, res.StatusCode(), res.Header(), res.Payload())
	switch {
	case errors.Is(err, runtime.ErrContentType):
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorResponse, OperationID: id, Err: err})
	case err != nil:
		a.fail(w, r, &runtime.HandlerError{Kind: runtime.ErrorService, OperationID: id, Err: err})
	}
}

// WithRouter registers the routes on r instead of a new ControllerRegister.
func WithRouter(r *web.ControllerRegister) ServerOption {
	return func(o *ServerOptions) {
		o.Router = r
	}
}

// NewRouter registers every operation on a beego ControllerRegister.
func NewRouter(svc PetsInterface, opts ...ServerOption) *web.ControllerRegister {
	o := NewServerOptions(opts...)
	a := NewHTTPAdapter(svc, opts...)
	wrap := func(h http.Handler) http.Handler {
		for i := len(o.Middleware) - 1; i >= 0; i-- {
			h = o.Middleware[i](h)
		}
		return h
	}
	register := func(r *web.ControllerRegister, route func(http.Handler) http.Handler) {
		r.AddMethod("GET", "/pets", handle(route(http.HandlerFunc(a.ListPets))))
		r.AddMethod("POST", "/pets", handle(route(http.HandlerFunc(a.CreatePet))))
		r.AddMethod("GET", "/pets/:id", handle(route(http.HandlerFunc(a.GetPet))))
		r.AddMethod("DELETE", "/pets/:id", handle(route(http.HandlerFunc(a.DeletePet))))
		r.AddMethod("GET", "/ping", handle(route(http.HandlerFunc(a.Ping))))
	}

	if r, _ := o.Router.(*web.ControllerRegister); r != nil {
		register(r, wrap)
		return r
	}
	r := web.NewControllerRegister()
	register(r, wrap)
	r.Any("/*", handle(wrap(http.NotFoundHandler())))
	return r
}

func handle(h http.Handler) web.HandleFunc {
	return func(c *bcontext.Context) {
		r := c.Request
		for key, value := range c.Input.Params() {
			if len(key) > 1 && key[0] == ':' {
				r.SetPathValue(key[1:], value)
			}
		}
		// beego has read a form body already, so it goes back into the body for h.
		if r.PostForm != nil && strings.HasPrefix(r.Header.Get("Content-Type"), "application/x-www-form-urlencoded") {
			r.Body = io.NopCloser(strings.NewReader(r.PostForm.Encode()))
		}
		h.ServeHTTP(c.ResponseWriter, r)
	}
}